enemy tries to win now
he plays badly, but he does aim to kil lth eplayer. all bulldozers are set to target the goal tree as soon as the spawn. they are on a navmesh, so they will path around hills ( if you change the terrain, re bake the navmesh)

fixed the ranged monkeys, projectiles work properly now.  the game gets harder as time goes on.

i left a hook in the health component (a comment anyway, where you can put the code to start the game over stuff.)